requirements.txt

PyGithub
openai
python-dotenv
requests
crewai-tools
